requirements.txt
accelerate==0.34.2
bitstring==4.2.3
clip-interrogator==0.6.0
datasets==2.15.0
diffusers==0.30.3
hydra-core==1.3.2
loguru==0.7.2
opencv-contrib-python==4.10.0.84
peft==0.13.0
piqa==1.3.2
scikit-image==0.24.0
scipy==1.13.1
torch==2.3.1+cu121
torchvision==0.18.1+cu121
wandb==0.18.3